Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240528/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.kt
@@ -5,15 +5,15 @@ class Main {
     private val br = BufferedReader(InputStreamReader(System.`in`))
     private lateinit var tokens : StringTokenizer
 
-    private var m = 0
+    private var m = 0L
     private var a = 0
     private var b = 0
 
-    private var min = 0 //  최소 찾는 횟수
-    private var max = 0     //  최대 찾는 횟수
+    private var min = 0L        //  최소 찾는 횟수
+    private var max = 0L     //  최대 찾는 횟수
 
     fun solve() {
-        m = br.readLine().toInt()
+        m = br.readLine().toLong()
 
         min = m + 1
 
@@ -24,8 +24,8 @@ class Main {
         for(target in a .. b) {
             val cnt = binarySearch(target)
 
-            min = min.coerceAtMost(cnt)
-            max = max.coerceAtLeast(cnt)
+            min = min.coerceAtMost(cnt.toLong())
+            max = max.coerceAtLeast(cnt.toLong())
         }
 
         println("$min $max")
@@ -33,14 +33,14 @@ class Main {
 
     //  정석적인 이진탐색으로 target을 찾기 위한 탐색 횟수
     private fun binarySearch(target : Int) : Int {
-        var start = 1
+        var start = 1L
         var end = m
         var cnt = 1
 
         while(start <= end) {
             val mid = (start + end) / 2
 
-            if(mid == target)
+            if(mid == target.toLong())
                 break
             else if(mid < target)
                 start = mid + 1
